viewer/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
pydantic-settings>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
